requirements.txt

streamlit
pycryptodome
